most of the results done for the poster
in results :
-in plots : we have the flop value plots but they are hard to read so instead
- in results/flops we have the table images for flops resuls (one for 32x32 and the other for 224x224 which also compares the original sfcnn) we can put these in the poster to show the difference in flops
- in overall_test_accuracy.txt, we have the results of our trained models on the cifar 10 test dataset.
-in test_results.jpg we have the table of the results i mentioned above
----------------------------------------
train_results.ipynb : it is directly inside the repository. inside, i plotted different performances of different models (compared all of them together, compared only the sfcnns and then compared the best sfcnn with other models) we can use some of them
---------------------------------------
to put in the final work more properly, i made 2 SFCNN_trainer files and one baselines file and one for convnext. the others are not necessary (when it comes to training). otherwise we should keep modules etc